+++ src/main/java/online/kingdomkeys/kingdomkeys/entity/magic/FaithBeamEntity.java
@@ -5,16 +5,19 @@ import net.minecraft.network.syncher.EntityDataSerializers;
 import net.minecraft.network.syncher.SynchedEntityData;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.MoverType;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.projectile.ThrowableProjectile;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.levelgen.Heightmap;
 import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.HitResult;
 import online.kingdomkeys.kingdomkeys.damagesource.KKDamageTypes;
+import online.kingdomkeys.kingdomkeys.data.PlayerData;
 import online.kingdomkeys.kingdomkeys.data.WorldData;
 import online.kingdomkeys.kingdomkeys.entity.ModEntities;
 import online.kingdomkeys.kingdomkeys.lib.Party;
+import online.kingdomkeys.kingdomkeys.lib.Strings;
+import online.kingdomkeys.kingdomkeys.util.Utils;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -59,7 +62,7 @@ public class FaithBeamEntity extends BaseMagicProjectile {
 			return;
 		}
 
-		if (!level().isClientSide && entityData.get(EXPANDING)) {
+		if (entityData.get(EXPANDING)) {
 			expandingTicks++;
 
 			orbitRadius += 0.8D + expandingTicks * 0.01D;
@@ -105,6 +108,25 @@ public class FaithBeamEntity extends BaseMagicProjectile {
 				continue;
 			}
 
+			if (hitTargets.isEmpty()) { //We heal here so it's not OP
+				float healAmount = 2;
+				if (getOwner() instanceof Player player) {
+					PlayerData playerData = PlayerData.get(player);
+					int localLevel = Utils.getMagicHighestLocalLevel(playerData.getEquippedMagics(), Strings.Magic_Faith, 0);
+
+					float totalHealPercent = 0.4F + localLevel * 0.1F; //0.5F --> 1.0F
+					int beamCount = 6;
+
+					switch (localLevel) {
+						case 3, 4 -> beamCount = 7;
+						case 5, 6 -> beamCount = 8;
+					}
+
+					healAmount = player.getMaxHealth() * totalHealPercent / beamCount;
+					float factor = 1F;
+					player.heal(healAmount * factor);
+				}
+			}
 			if (hitTargets.contains(target.getUUID())) {
 				continue;
 			}